scripts/ai-bot/requirements.txt
#!/usr/bin/env python3
"""
Requirements file for AI Bot dependencies
Install with: pip install -r requirements.txt
"""

# YAML处理
PyYAML>=6.0

# HTTP请求
requests>=2.31

# HTML解析
beautifulsoup4>=4.12.0

# 图像处理（用于未来的OCR功能）
Pillow>=10.0

# 数据验证
pydantic>=2.0

# GitHub API
PyGithub>=2.0.0

# URL解析和日期处理
python-dateutil>=2.8

# Pydantic配置
pydantic-settings>=2.0.0


# 可选：OCR支持（如后续需要图片文字提取）
# pytesseract>=0.3
# opencv-python>=4.8
